HSEARCH-2849 Leverage content-length computation when computing AWS hash
